3.Server/pyproject.toml
[tool.poetry]
name = "whatsderserver"
version = "0.1.0"
description = ""
authors = ["FengJing Liu <sdsafaflfj814@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


fastapi = "^0.111.0"
baidu-aip = "^4.16.13"
requests = "^2.32.3"
openai = "^1.34.0"
pillow = "^10.3.0"
numpy = "^2.0.0"
websocket-client = "^1.8.0"
ipykernel = "^6.29.4"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"